// pj-client/src/app/domain/dto/ConfiguracionRequest.dto.ts
export interface EspecialidadRequest{
    especialidad: string;
    app_user: string;
}

export interface SedeRequest{
    sede: string;
    app_user: string;
}

export interface TipoDocumentoRequest{
    tipo_documento: string;
    app_user: string;
}

export interface JuzgadoRequest{
    juzgado: string;
    app_user: string;
}

export interface MateriaRequest{
    materia: string;
    app_user: string;
}

export interface TipoProcesoRequest{
    tipo_proceso: string;
    app_user: string;
}

export interface GeneralRequest{
    institucion: string;
    direccion: string;
    ruc: string;
    departamento: string;
    app_user: string;
}